requirements.txt
acres==0.5.0
amqp==5.3.1
asgiref==3.9.0
billiard==4.2.1
black==25.1.0
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
Django==5.2.4
docx==0.2.4
etelemetry==0.3.1
filelock==3.18.0
fsspec==2025.5.1
httplib2==0.22.0
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.6
kombu==5.5.4
looseversion==1.3.0
lxml==6.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
pathspec==0.12.1
pdf2image==1.17.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
prov==2.1.1
psycopg2-binary==2.9.10
puremagic==1.30
pydot==4.0.1
PyMuPDF==1.26.3
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-pptx==1.0.2
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
redis==6.2.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scipy==1.16.0
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
traits==7.0.2
transformers==4.53.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
xlsxwriter==3.2.5
